// packages/server/src/resolvers/types.ts
export type ID = string

/**
 * This enum is structurally compatible with the enum generated by
 * graphql-codegen.
 */
export enum Episode {
  Newhope = "NEWHOPE",
  Empire = "EMPIRE",
  Jedi = "JEDI"
}

export interface Character {
  id: ID
  name?: string
  friends: ID[]
  appearsIn: Episode[]
  secretBackstory?: string
}

export interface Droid extends Character {
  primaryFunction?: string
}

export interface Human extends Character {
  homePlanet?: string
}